% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rgaussjordan-package.R
\name{cpp4rgaussjordan-package}
\alias{cpp4rgaussjordan-package}
\title{A Naive Implementation of the Gauss-Jordan Algorithm}
\description{
Just to demonstrate the vendoring capabilities of the 'cpp4r'
 package. This is a very simple implementation to show code syntax.
}

