% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rgaussjordan-package.R
\name{invert_matrix}
\alias{invert_matrix}
\title{Invert (some) square matrices}
\usage{
invert_matrix(A)
}

\arguments{
\item{A}{numeric matrix}
}

\value{
numeric matrix
}

\description{
Invert (some) square matrices
}

\examples{
A <- matrix(c(2, 1, 3, -1), nrow = 2, ncol = 2)
invert_matrix(A)
}

