% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rgaussjordan-package.R
\name{solve_system}
\alias{solve_system}
\title{Solve (some) linear systems}
\usage{
solve_system(A, b)
}

\arguments{
\item{A}{numeric matrix}

\item{b}{numeric matrix}
}

\value{
numeric matrix
}

\description{
Solve (some) linear systems
}

\examples{
A <- matrix(c(2, 1, 3, -1), nrow = 2, ncol = 2)
b <- matrix(c(7, 4), nrow = 2, ncol = 1)
solve_system(A, b)
}

